feat: add more features
I wanted to breakdown and add features one by one in the commits but as
I started working one thing led to another and a plethora of features were
added and tested and it became one big huge commit.
Now a good functioning app exists at this stage so I thought it is at better
to make at commit here.