feat: add confirmOnExit feature
This is a feature I usually add to all my apps. This is becoming somewhat
of happy ritual to add this feature at end of all required functionality.
I feel the App turned out to be pretty good.
Should add readme soon with all the screenshots and examples which might be
useful for others who may check this project out later.